Delete android-ktx module and move its content in androidMain source set
