--- android-ktx/build.gradle
@@ -1,64 +1,0 @@
-plugins {
-    id 'com.android.library'
-    id 'kotlin-android'
-    id 'kotlin-platform-android'
-    id 'digital.wup.android-maven-publish' version '3.6.2'
-    id 'mirego.release' version '2.0'
-    id 'mirego.publish' version '1.0'
-}
-
-repositories {
-    google()
-    mavenCentral()
-    mavenLocal()
-    maven { url "https://dl.bintray.com/kotlin/ktor" }
-    maven { url "https://kotlin.bintray.com/kotlinx" }
-    maven { url 'https://jitpack.io' }
-    maven { url('https://s3.amazonaws.com/mirego-maven/public') }
-}
-
-group 'com.mirego.trikot.http'
-
-android {
-    compileSdkVersion 30
-
-    defaultConfig {
-        minSdkVersion 14
-        targetSdkVersion 30
-    }
-}
-
-configurations.all {
-    exclude group: "org.reactivestreams"
-}
-
-dependencies {
-    api "com.mirego.trikot:trikotFoundation:$trikot_foundation_version"
-    api "com.mirego.trikot:streams:$trikot_streams_version"
-    api 'com.mirego.trikot:http:1.0.7'
-    implementation "io.ktor:ktor-client-android:$ktor_version"
-    api "io.ktor:ktor-client-logging-jvm:$ktor_version"
-    implementation 'androidx.lifecycle:lifecycle-extensions:2.2.0'
-    implementation 'androidx.lifecycle:lifecycle-reactivestreams-ktx:2.3.0'
-}
-
-release {
-    checkTasks = ['check']
-    buildTasks = ['publish']
-    updateVersionPart = 2
-    tagPrefix = 'android-ktx-'
-}
-
-task sourcesJar(type: Jar) {
-    from android.sourceSets.main.java.srcDirs
-    archiveClassifier.set('sources')
-}
-
-publishing {
-    publications {
-        mavenAar(MavenPublication) {
-            from components.android
-            artifact sourcesJar
-        }
-    }
-}
