--- android-ktx/gradle.properties
@@ -1,2 +1,0 @@
-#Fri Dec 04 11:38:09 EST 2020
-version=2.0.0-SNAPSHOT
