--- android-ktx/src/main/AndroidManifest.xml
@@ -1,4 +1,0 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-          package="com.mirego.trikot.http.android.ktx">
-    <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
-</manifest>
