+++ http/build.gradle
@@ -21,6 +21,7 @@ repositories {
     mavenLocal()
     maven { url "https://kotlin.bintray.com/kotlinx" }
     maven { url 'https://jitpack.io' }
+    maven { url 'https://plugins.gradle.org/m2/' }
     maven { url('https://s3.amazonaws.com/mirego-maven/public') }
 }
 
@@ -64,6 +65,12 @@ kotlin {
 
         androidMain {
             dependsOn commonMain
+            dependencies {
+                api "io.ktor:ktor-client-logging-jvm:$ktor_version"
+                implementation "io.ktor:ktor-client-android:$ktor_version"
+                implementation 'androidx.lifecycle:lifecycle-extensions:2.2.0'
+                implementation 'androidx.lifecycle:lifecycle-reactivestreams-ktx:2.3.0'
+            }
         }
 
         nativeMain {
