+++ http/src/commonMain/kotlin/com/mirego/trikot/http/HttpRequest.kt
@@ -1,8 +1,8 @@
 package com.mirego.trikot.http
 
 import com.mirego.trikot.streams.cancellable.CancellableManager
-import kotlin.js.JsName
 import org.reactivestreams.Publisher
+import kotlin.js.JsName
 
 interface HttpRequest {
     @JsName("execute")
