+++ settings.gradle
@@ -22,4 +22,4 @@ pluginManagement {
 }
 rootProject.name = 'trikot.http'
 
-include ':http', ':android-ktx'
+include ':http'
